mailService/requirements.txt
Flask
requests
flasgger
python-dotenv
gunicorn
flask_cors
pymongo
